move BiddingNavigator to Navigation
