+++ src/app/Components/Bidding/Components/PaymentInfo.tsx
@@ -4,7 +4,7 @@ import { Token } from "@stripe/stripe-react-native"
 import { Card } from "@stripe/stripe-react-native/lib/typescript/src/types/Token"
 import { FlexProps } from "app/Components/Bidding/Elements/Flex"
 import { Address, PaymentCardTextFieldParams } from "app/Components/Bidding/types"
-import { BiddingNavigationStackParams } from "app/Components/Containers/BiddingNavigator"
+import { BiddingNavigationStackParams } from "app/Navigation/AuthenticatedRoutes/BiddingNavigator"
 import React from "react"
 import { View } from "react-native"
 import { BidInfoRow } from "./BidInfoRow"
