+++ src/app/Components/Bidding/Screens/BidResult.tsx
@@ -3,9 +3,9 @@ import { StackActions } from "@react-navigation/native"
 import { NativeStackScreenProps } from "@react-navigation/native-stack"
 import { Timer } from "app/Components/Bidding/Components/Timer"
 import { BidFlowContextStore } from "app/Components/Bidding/Context/BidFlowContextProvider"
-import { BiddingNavigationStackParams } from "app/Components/Containers/BiddingNavigator"
 import { Markdown } from "app/Components/Markdown"
 import { NavigationHeader } from "app/Components/NavigationHeader"
+import { BiddingNavigationStackParams } from "app/Navigation/AuthenticatedRoutes/BiddingNavigator"
 import { unsafe__getEnvironment } from "app/store/GlobalStore"
 import { dismissModal, navigate } from "app/system/navigation/navigate"
 import { useBackHandler } from "app/utils/hooks/useBackHandler"
