+++ src/app/Components/Bidding/Screens/ConfirmBid/index.tsx
@@ -16,10 +16,10 @@ import { BidFlowContextStore } from "app/Components/Bidding/Context/BidFlowConte
 import { Flex } from "app/Components/Bidding/Elements/Flex"
 import { bidderPositionQuery } from "app/Components/Bidding/Screens/ConfirmBid/BidderPositionQuery"
 import { PriceSummary } from "app/Components/Bidding/Screens/ConfirmBid/PriceSummary"
-import { BiddingNavigationStackParams } from "app/Components/Containers/BiddingNavigator"
 import { Modal } from "app/Components/Modal"
 import { NavigationHeader } from "app/Components/NavigationHeader"
 import { LegacyNativeModules } from "app/NativeModules/LegacyNativeModules"
+import { BiddingNavigationStackParams } from "app/Navigation/AuthenticatedRoutes/BiddingNavigator"
 import { partnerName } from "app/Scenes/Artwork/Components/ArtworkExtraLinks/partnerName"
 import { navigate } from "app/system/navigation/navigate"
 import { AuctionWebsocketContextProvider } from "app/utils/Websockets/auctions/AuctionSocketContext"
