+++ src/app/Components/Bidding/Screens/CreditCardForm.tsx
@@ -10,10 +10,10 @@ import {
 import { findCountryNameByCountryCode } from "app/Components/Bidding/Utils/findCountryNameByCountryCode"
 import { creditCardFormValidationSchema } from "app/Components/Bidding/Validators/creditCardFormFieldsValidationSchema"
 import { Address } from "app/Components/Bidding/types"
-import { BiddingNavigationStackParams } from "app/Components/Containers/BiddingNavigator"
 import { CountrySelect } from "app/Components/CountrySelect"
 import { CreditCardField } from "app/Components/CreditCardField/CreditCardField"
 import { NavigationHeader } from "app/Components/NavigationHeader"
+import { BiddingNavigationStackParams } from "app/Navigation/AuthenticatedRoutes/BiddingNavigator"
 import { useFormik } from "formik"
 import { memo, useCallback, useRef } from "react"
 import { KeyboardAvoidingView, ScrollView } from "react-native"
