+++ src/app/Components/Bidding/Screens/PhoneNumberForm.tsx
@@ -1,10 +1,10 @@
 import { Box, Button } from "@artsy/palette-mobile"
 import { NativeStackScreenProps } from "@react-navigation/native-stack"
 import { BottomAlignedButtonWrapper } from "app/Components/Buttons/BottomAlignedButtonWrapper"
-import { BiddingNavigationStackParams } from "app/Components/Containers/BiddingNavigator"
 import { Input } from "app/Components/Input"
 import { PhoneInput } from "app/Components/Input/PhoneInput"
 import { NavigationHeader } from "app/Components/NavigationHeader"
+import { BiddingNavigationStackParams } from "app/Navigation/AuthenticatedRoutes/BiddingNavigator"
 import { ProvideScreenTracking, Schema, track } from "app/utils/track"
 import { PageNames } from "app/utils/track/schema"
 import { useEffect, useRef, useState } from "react"
