+++ src/app/Components/Bidding/Screens/RegistrationResult.tsx
@@ -3,9 +3,9 @@ import { NativeStackScreenProps } from "@react-navigation/native-stack"
 import { Icon20 } from "app/Components/Bidding/Components/Icon"
 import { Title } from "app/Components/Bidding/Components/Title"
 import { Flex } from "app/Components/Bidding/Elements/Flex"
-import { BiddingNavigationStackParams } from "app/Components/Containers/BiddingNavigator"
 import { Markdown } from "app/Components/Markdown"
 import { NavigationHeader } from "app/Components/NavigationHeader"
+import { BiddingNavigationStackParams } from "app/Navigation/AuthenticatedRoutes/BiddingNavigator"
 import { dismissModal } from "app/system/navigation/navigate"
 import { defaultRules } from "app/utils/renderMarkdown"
 import { Schema, screenTrack } from "app/utils/track"
