+++ src/app/Components/Bidding/Screens/SelectMaxBid.tsx
@@ -4,9 +4,9 @@ import { SelectMaxBidQuery } from "__generated__/SelectMaxBidQuery.graphql"
 import { SelectMaxBid_me$key } from "__generated__/SelectMaxBid_me.graphql"
 import { SelectMaxBid_saleArtwork$key } from "__generated__/SelectMaxBid_saleArtwork.graphql"
 import { BidFlowContextStore } from "app/Components/Bidding/Context/BidFlowContextProvider"
-import { BiddingNavigationStackParams } from "app/Components/Containers/BiddingNavigator"
 import { NavigationHeader } from "app/Components/NavigationHeader"
 import { Select } from "app/Components/Select"
+import { BiddingNavigationStackParams } from "app/Navigation/AuthenticatedRoutes/BiddingNavigator"
 import { dismissModal } from "app/system/navigation/navigate"
 
 import { NoFallback, SpinnerFallback, withSuspense } from "app/utils/hooks/withSuspense"
