+++ src/app/Components/Containers/RegistrationFlow.tsx
@@ -1,6 +1,6 @@
 import { Screen } from "@artsy/palette-mobile"
 import { TimeOffsetProvider } from "app/Components/Bidding/Context/TimeOffsetProvider"
-import { BiddingNavigator } from "app/Components/Containers/BiddingNavigator"
+import { BiddingNavigator } from "app/Navigation/AuthenticatedRoutes/BiddingNavigator"
 
 export const RegistrationFlow: React.FC<{ saleID: string }> = (props) => {
   return (
