+++ src/app/Navigation/AuthenticatedRoutes/BiddingNavigator.tsx
